+++ LogsManager/Classes/Log.swift
@@ -137,17 +137,15 @@ public func logVerbose(_ message: @autoclosure () -> String,
 }
 
 /// Trace log function. This one is to trace execution path. Usually, the log should be placed as the first line in the method or function.
-/// - parameter message: Message to log.
 /// - parameter logComponents: Components this log belongs to, e.g. `.network`, `.keychain`, ... . Autodetect if `nil`.
-/// - parameter data: Additional data to log.
 public func logTrace(logComponents: [LogComponent]? = nil,
                      file: String = #file,
                      function: String = #function,
                      line: UInt = #line) {
     
-    logMessage("\(file):\(line) \(function)",
+    logMessage("\(String.getFileName(filePath: file)):\(line) \(function)",
                logComponents: logComponents,
-               flag: .data,
+               flag: .trace,
                data: nil,
                file: file,
                function: function,