+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 All notable changes to this project will be documented in this file.
 `LogsManager` adheres to [Semantic Versioning](http://semver.org/).
 
+## [12.3.1](https://github.com/APUtils/LogsManager/releases/tag/12.3.1)
+Released on `2023-06-21
+
+#### Added
+- Trace logs for `RoutableLogger`
+
+#### Changed
+- Output file name only instead of whole path for trace log
+
+#### Fixed
+- Trace log level fix
+
+
 ## [12.3.0](https://github.com/APUtils/LogsManager/releases/tag/12.3.0)
 Released on `2023-06-20